Ostalo/ObstojecaKoda/IPAD/requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
idna==2.8
joblib==0.14.0
jsonpatch==1.24
jsonpointer==2.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
pprint==0.1
pyparsing==2.4.2
python-dateutil==2.8.0
pyzmq==18.1.1
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
torch==1.3.0
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.4.1
tornado==6.0.3
tqdm==4.36.1
urllib3==1.25.7
visdom==0.1.8.9
websocket-client==0.56.0
